Add files via upload
This is the code. It works only in python3 with matplotlib.pyplot installed. It keeps asking you for a seed until you enter'0000'.
I know it works in IDLE, but some editors like Visual Studio might not work. If you have a request for me, or an issue with my code, feel free to leave a comment or issue.